+++ androidx-kt/src/androidTest/java/me/panpf/androidxkt/test/app/PermissionTest.kt
@@ -21,7 +21,7 @@ import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
 import me.panpf.androidxkt.app.isGrantPermission
 import me.panpf.androidxkt.app.isGrantPermissions
-import me.panpf.androidxkt.os.isAtLeastM
+import me.panpf.androidxkt.isAtLeastM
 import org.junit.Assert
 import org.junit.Test
 import org.junit.runner.RunWith