+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * 升级 Javax 依赖
 * Color 增加 argbEvaluate 方法
 * 增加 WeakAsyncTask
+* Build、Process、System、Thread 合并到 Android 中
 
 ## v0.1
 * 初次发布